chore(mirror): Make server modules an array
Before there was a main and modules. For simplicity use a single array.
