+++ mirror/mirror-schema/src/server.ts
@@ -2,16 +2,27 @@ import * as v from 'shared/src/valita.js';
 import {firestoreDataConverter} from './converter.js';
 import * as path from './path.js';
 
+export const moduleSchema = v.object({
+  name: v.string(),
+  // filename is the filename used in Google Cloud Storage. It has a GUID in it.
+  filename: v.string(),
+  type: v.union(v.literal('esm'), v.literal('text')),
+});
+
+export type Module = v.Infer<typeof moduleSchema>;
+
 export const releaseChannelSchema = v.union(
   v.literal('canary'),
   v.literal('stable'),
 );
 
+export type ReleaseChannel = v.Infer<typeof releaseChannelSchema>;
+
 export const serverSchema = v.object({
   major: v.number(),
   minor: v.number(),
   patch: v.number(),
-  module: v.string(),
+  modules: v.array(moduleSchema),
   channel: releaseChannelSchema,
 });
 
@@ -21,6 +32,6 @@ export const serverDataConverter = firestoreDataConverter(serverSchema);
 
 export const SERVER_COLLECTION = 'servers';
 
-export function serverPath(serverID: string): string {
-  return path.join(SERVER_COLLECTION, serverID);
+export function serverPath(version: string): string {
+  return path.join(SERVER_COLLECTION, version);
 }
